+++ src/js/qiniu.js
@@ -1,2 +1,0 @@
-var qiniu = require('qiniu-js');
-